+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### General
 
+### Bug Fixes
+
 ## [0.2.10](https://github.com/Aleph-Alpha-Research/eval-framework/compare/v0.2.9...v0.2.10) (2026-01-27)